requirements/requirements-memory_gym.txt
# This file was autogenerated by uv via the following command:
#    uv export --no-hashes --output-file ../../requirements/requirements-memory_gym.txt
-e .
absl-py==2.3.1
    # via tensorboard
appdirs==1.4.4
    # via wandb
backports-shutil-get-terminal-size==1.0.0
    # via reprint
certifi==2025.6.15
    # via
    #   requests
    #   sentry-sdk
charset-normalizer==3.4.2
    # via requests
click==8.2.1
    # via wandb
cloudpickle==3.1.1
    # via gymnasium
colorama==0.4.6
    # via
    #   click
    #   reprint
    #   tyro
docker-pycreds==0.4.0
    # via wandb
docstring-parser==0.16
    # via tyro
einops==0.7.0
    # via ppo-trxl
farama-notifications==0.0.4
    # via gymnasium
filelock==3.18.0
    # via torch
fsspec==2025.5.1
    # via torch
gitdb==4.0.12
    # via gitpython
gitpython==3.1.44
    # via wandb
grpcio==1.66.2
    # via tensorboard
gymnasium==0.29.0
    # via
    #   memory-gym
    #   minigrid
idna==3.10
    # via requests
jinja2==3.1.6
    # via torch
markdown==3.8.2
    # via tensorboard
markdown-it-py==3.0.0
    # via rich
markupsafe==3.0.2
    # via
    #   jinja2
    #   werkzeug
mdurl==0.1.2
    # via markdown-it-py
memory-gym==1.0.2
    # via ppo-trxl
minigrid==2.5.0
    # via ppo-trxl
mpmath==1.3.0
    # via sympy
networkx==3.4.2 ; python_full_version < '3.11'
    # via torch
networkx==3.5 ; python_full_version >= '3.11'
    # via torch
numpy==2.2.6 ; python_full_version < '3.11'
    # via
    #   gymnasium
    #   minigrid
    #   opencv-python
    #   tensorboard
numpy==2.3.1 ; python_full_version >= '3.11'
    # via
    #   gymnasium
    #   minigrid
    #   opencv-python
    #   tensorboard
nvidia-cublas-cu11==11.11.3.6 ; platform_machine == 'x86_64' and sys_platform == 'linux'
    # via
    #   nvidia-cudnn-cu11
    #   nvidia-cusolver-cu11
    #   torch
nvidia-cuda-cupti-cu11==11.8.87 ; platform_machine == 'x86_64' and sys_platform == 'linux'
    # via torch
nvidia-cuda-nvrtc-cu11==11.8.89 ; platform_machine == 'x86_64' and sys_platform == 'linux'
    # via torch
nvidia-cuda-runtime-cu11==11.8.89 ; platform_machine == 'x86_64' and sys_platform == 'linux'
    # via torch
nvidia-cudnn-cu11==9.1.0.70 ; platform_machine == 'x86_64' and sys_platform == 'linux'
    # via torch
nvidia-cufft-cu11==10.9.0.58 ; platform_machine == 'x86_64' and sys_platform == 'linux'
    # via torch
nvidia-curand-cu11==10.3.0.86 ; platform_machine == 'x86_64' and sys_platform == 'linux'
    # via torch
nvidia-cusolver-cu11==11.4.1.48 ; platform_machine == 'x86_64' and sys_platform == 'linux'
    # via torch
nvidia-cusparse-cu11==11.7.5.86 ; platform_machine == 'x86_64' and sys_platform == 'linux'
    # via torch
nvidia-nccl-cu11==2.21.5 ; platform_machine == 'x86_64' and sys_platform == 'linux'
    # via torch
nvidia-nvtx-cu11==11.8.86 ; platform_machine == 'x86_64' and sys_platform == 'linux'
    # via torch
opencv-python==4.11.0.86
    # via ppo-trxl
packaging==25.0
    # via tensorboard
protobuf==4.21.0
    # via
    #   tensorboard
    #   wandb
psutil==7.0.0
    # via wandb
pygame==2.4.0
    # via
    #   memory-gym
    #   minigrid
pygments==2.19.2
    # via rich
pyyaml==6.0.2
    # via wandb
reprint==0.6.0
    # via ppo-trxl
requests==2.32.4
    # via wandb
rich==14.0.0
    # via tyro
sentry-sdk==2.32.0
    # via wandb
setproctitle==1.3.6
    # via wandb
setuptools==80.9.0
    # via
    #   tensorboard
    #   torch
    #   triton
    #   wandb
shtab==1.7.2
    # via tyro
six==1.17.0
    # via
    #   docker-pycreds
    #   reprint
    #   tensorboard
smmap==5.0.2
    # via gitdb
sympy==1.14.0
    # via torch
tensorboard==2.19.0
    # via ppo-trxl
tensorboard-data-server==0.7.2
    # via tensorboard
torch==2.7.1+cu118
    # via
    #   ppo-trxl
    #   torchaudio
torchaudio==2.7.1+cu118
    # via ppo-trxl
triton==3.3.1 ; platform_machine == 'x86_64' and sys_platform == 'linux'
    # via torch
typing-extensions==4.14.1
    # via
    #   gymnasium
    #   rich
    #   torch
    #   tyro
tyro==0.8.14
    # via ppo-trxl
urllib3==2.5.0
    # via
    #   requests
    #   sentry-sdk
wandb==0.16.6
    # via ppo-trxl
werkzeug==3.1.3
    # via tensorboard
